requirements.txt
lmdb==0.94
opencv-python==3.4.3.18
Pillow==5.1.0
numpy==1.15.1
requests==2.20.0
beautifulsoup4==4.6.0
tqdm==4.23.4
torchvision==0.2.1
http://download.pytorch.org/whl/cpu/torch-0.4.1-cp36-cp36m-linux_x86_64.whl ; sys_platform == "linux"
